64 bit entity ids

+++ tinyecs/meta.hpp
@@ -20,7 +20,7 @@ namespace tinyecs {
 
 	using size_type = size_t;
 	using type_index = uint32_t;
-	using entity = uint32_t;
+	using entity = uint64_t;
 	using component_id = type_index;
 
 	constexpr entity null_entity = entity(0);
